refactor(hub-common): make setContentExtent() and setContentBoundary() internal only
affects: @esri/hub-common
